+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [UNRELEASED]
+
+### Fixed
+
+- Applied various improvements to logging.
+
 ## [3.1.1] - 2024-07-15
 
 ### Fixed